+++ changelog.md
@@ -1,2 +1,3 @@
 ## Version 3.2
 - Fixed bug with entities being misplaced after map size is expanded.
+- Made currently open file show up in the window title.